+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added GP Stroke From Spline node.
 - Added Spline From GP Stroke node.
+- Added Offset GP Stroke node.
 
 ### Fixed